% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow.MVITR.R
\name{grow.MVITR}
\alias{grow.MVITR}
\title{Grows a large interaction tree}
\usage{
\method{grow}{MVITR}(data.in, split.var.in, test.in = NULL,
  min.ndsz.in = 50, n0.in = 5, ctg.in = NULL, max.depth.in = 15,
  mtry.in = length(split.var.in), AIPWE.in = FALSE,
  standardize.ae = FALSE, standardize.surv = FALSE, stabilize.ae = TRUE,
  stabilize.surv = TRUE, n.folds = 10, stabilize.type.in = c("rf",
  "linear"))
}
\arguments{
\item{data.in}{data set from which the tree is to be grown.  Must contain outcome, binary 
treatment indicator, columns of splitting covariates, and column of probability of being
in treatment group.}

\item{split.var.in}{columns of potential spliting variables. Required input.}

\item{test.in}{testing data}

\item{min.ndsz.in}{minimum number of observations required to call a node terminal. Defaults to 20.}

\item{n0.in}{minimum number of treatment/control observations needed in a split to call a node terminal. Defaults to 5.}

\item{ctg.in}{identifies the categorical input columns.  Defaults to NULL.  Not available yet.}

\item{max.depth.in}{controls the maximum depth of the tree. Defaults to 15.}

\item{AIPWE.in}{logical. Should AIPWE (TRUE) or IPWE (FALSE) be used.}

\item{stabilize.type.in}{gives the method used for calculating residuals. Current options are 'rf' for random forest and 'linear' for linear model.}
}
\value{
Summary of a single interaction tree. Each `node` begins with "0" indicating the root node, 
followed by a "1" or "2" indicating the less than (or left) child node or greater than (or right) child node. 
Additionally, the number of observations `size`, number treated `n.1`, number on control `n.0`, and treatment effect `trt.effect`
summaries are provided.  The splitting information includes the column of the chosen splitting variable `var`, the variable name 'vname',
the direction the treatment is sent `cut.1` ("r" for right child node, and "l" for left), the chosen split value `cut.2`, 
and the estimated value function `score`.
}
\description{
This function grows an interaction tree using either the IPWE or AIPWE method (AIPWE=F, T).
}
\examples{
dat <- gdataM(n=1000, depth=2, beta1=3, beta2=1)
# Generates tree using simualated EMR data with splitting variables located in columns 1-4.
tree <- grow.ITR(data=dat, split.var=1:4)
}
