% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haoda.partition.ITR.R
\name{haoda.partition.ITR}
\alias{haoda.partition.ITR}
\title{Generates partition summary based on itr value.}
\usage{
haoda.partition.ITR(dat, test = NULL, name = "0", min.ndsz = 20, n0 = 5,
  split.var, outcome = c("time"), ctg = ctg, max.depth = 15,
  mtry = length(split.var), dat.rest = NULL, max.score = NULL,
  AIPWE = AIPWE, haoda.ae.level = NA, use.other.nodes = TRUE)
}
\arguments{
\item{dat}{data set from which the partition is to be made.  Must contain outcome, binary 
treatment indicator, columns of splitting covariates, and column of probability of being
in treatment group.}

\item{test}{testing data set}

\item{name}{internal variable the keeps track of the current node(s).}

\item{min.ndsz}{minimum number of observations required to call a node terminal. Defaults to 20.}

\item{n0}{minimum number of treatment/control observations needed in a split to call a node terminal. Defaults to 5.}

\item{split.var}{columns of potential spliting variables. Required input.}

\item{ctg}{identifies the categorical input columns.  Defaults to NULL.  Not available yet.}

\item{max.depth}{controls the maximum depth of the tree. Defaults to 15.}

\item{mtry}{sets the number of randomly selected splitting variables to be included. Defaults to number of splitting variables.}

\item{dat.rest}{data outside the current node being split}

\item{max.score}{controls the minimum score required to make an additional split (internally controlled).}

\item{AIPWE}{indicator for AIPWE estimation}

\item{haoda.ae.level}{cutoff for acceptable risk in a given split (numeric)}
}
\value{
summary of the best split for a given data frame.
}
\description{
The partitioning function is used inside of the tree growing functions.  It 
selects the best split among a set of predictors based on the IPWE or AIPWE value.
}
