% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvdatAlt.R
\name{mvdatAlt}
\alias{mvdatAlt}
\title{Simulates data from an RCT design}
\usage{
mvdatAlt(n = 500, K = 50, K.E = 10, design = c("emr", "rct"),
  cut.type = c("linear", "single"), beta11 = 1, beta12 = 1, beta21 = 1,
  rateC = 0.1, StudyPeriod = 36, n.noise.cov = 10, cut1 = 0.5,
  cut2 = 0.3, cut3 = 0.1, depth1 = 1, depth2 = 1, offset1 = 50,
  lambda = 10, rho = 3, dispersion = 20, sigma = 0.5)
}
\arguments{
\item{n}{size of the dataset to be generate.  Defaults to 100.}

\item{K}{internal variable used to generate the fineness of the unit interval on which covariates are generated. Defaults to 50.}

\item{K.E}{internal variable used to generate the fineness of the unit interval on which noise covariates are generated. Defaults to 10.}

\item{cut.type}{Indicates the type of subgroup structure (linear or single)}

\item{beta11}{controls the strength of the treatment effect for survival outcome in treatment group. Defaults to 0.5.}

\item{beta12}{controls the strength of the control effect for survival outcome in control group. Defaults to 0.5.}

\item{beta21}{controls the strength of the control effect for AE outcome. Defaults to 1.}

\item{rateC}{censoring rate. Defaults to 0.1.}

\item{StudyPeriod}{maximum observation period. Defaults to 36. Not currently in use.}

\item{n.noise.cov}{number of noise covariates to generate. Defaults to 10.}

\item{cut1, cut2, cut3}{controls where the cutpoints are to define subgroups.}

\item{depth1}{gives the number of interacting covariates. If set to 1, then 
then covariate X1 interacts with treatment. If set to another value, then 
covariates X1 and X3 both interact with treatment effect (one-way interactions). Defaults to 1.}

\item{depth2}{Controls the number of interacting covariates in the AE outcome. Defaults to 1.}

\item{offset1}{The multiplier of the time scale in survival estimates}

\item{lambda}{Scale parameter for the survival times}

\item{rho}{Shape parameter for the survival times}

\item{dispersion}{Dispersion parameter (size = dispersion) for NB dist}

\item{sigma}{controls standard deviation of random variation.  Defaults to 1.}
}
\value{
data frame containing y (outcome), x1-x4 (covariates), trt (treatment), prtx (probability of being in treatment group), and ID variable
}
\description{
Simulates data from an RCT according to the following model:
2 + 2*sign(x1<cut2) + beta1*trt*subgrp + beta2*(1-trt)*subgrp + N(0,sigma)
If depth=1, then subgrp=(x1<=cut1)
If depth!=1 then subgrp=(x1>=0.3 & x3>=0.1)
}
\examples{

# This generates a dataframe with 500 observations, X1 as the only variable interacting with 
# the treatment, and a signal to noise ratio of 2/2=1.
data<-rdat(n=500)
}
