\name{mvITR-package}
\alias{mvITR-package}
\alias{mvITR}
\docType{package}
\title{
  Discovery of multivariate ITR.
}
\description{
  This package can be used for discovery of risk controlled individualized treatment rules (ITRs) via risk controlled decision trees (rcDT) or risk controlled random forests (rcRF).
}
\details{
  Use the grow.ITR() and Build.RF.ITR() functions for risk controITR discovery. 
}
\author{
Kevin Doubleday, kdoub5ha@gmail.com.

Maintainer: Kevin Doubleday <kdoub5ha@gmail.com>
}
\references{
  NA. 
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
