% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ITR.R
\name{predict.ITR}
\alias{predict.ITR}
\title{Treatment Prediction for rcDT and rcRF Models}
\usage{
\method{predict}{ITR}(fit, new.data, split.var, ctgs = NULL)
}
\arguments{
\item{fit}{tree or forest object from `grow.ITR` or `Build.RF.ITR`.}

\item{new.data}{data for which predictions are desired}

\item{split.var}{splitting variables from the model fit}

\item{ctgs}{columns of categorical variables.}
}
\value{
A summary list of the following elements:

\item{SummaryTreat}{proportion of trees voting for treatment (trt=1). 
If input is rcDT (single tree) then SummaryTreat is a single number. 
If input is rcRF (forest) then SummaryTreat is a vector equal to the length of the number of trees.}

\item{trt.pred}{vector of treatment assignments {0, 1} based on the tree vote (single tree) or majority of tree votes (forest). This vector has length equal to the number of rows in `new.data`.}

\item{n.trees}{number of tree in `fit`}

\item{tree.votes}{matrix of votes for each tree for each subject in `new.data`. Rows correspond to trees in `fit` and columns correspond to subjects in `new.dat`.}

\item{data}{input data frame `new.data`}

\item{NA.trees}{number of trees returning no votes. In a forest, this is the number of null trees.}
}
\description{
Used to make treatment prediction for a rcDT and rcRF models. If the 
input is rcRF (forest), then the proportion of trees voting for treatment (`trt=1`) is returned. 
If the input is rcDT (single tree), then the function returns the vote (0 / 1) for the model.
}
\examples{
dat <- generateData()
# Generates rcDT using simualated data with splitting variables located in columns 1-10.
tree <- grow.ITR(data = dat, split.var = 1:10, 
                 risk.control = TRUE, risk.threshold = 2.75, 
                 lambda = 1)
# Predict treatment assignments for 1000 observations in `dat`
preds <- predict.ITR(fit = tree, new.data = dat, split.var = 1:10)
}
