\name{mvITR-package}
\alias{mvITR-package}
\alias{mvITR}
\docType{package}
\title{
  Discovery of multivariate ITR.
}
\description{
  This package can be used for discovery of multivariate ITR using both continuous and survival outcomes. Two main functions are provided which discover ITR via a conditional splitting algorithm or a sequential splitting algorithm.
}
\details{
  Use the grow.ITR() and grow.MVITR() functions for ITR discovery. 
}
\author{
Kevin Doubleday, kdoub5ha@gmail.com.

Maintainer: Kevin Doubleday <kdoub5ha@gmail.com>
}
\references{
  NA. 
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
